requirements.txt
# BrownButter Shopify Automation - Python Dependencies
# Install with: pip install -r requirements.txt

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
gspread>=5.11.0
oauth2client>=4.1.3

# HTTP Requests
requests>=2.31.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0.1

# Image Processing
Pillow>=10.0.0

# Cloudinary
cloudinary>=1.36.0

# Progress Bars
tqdm>=4.66.0

# Utilities
python-dateutil>=2.8.2

Flask==3.0.0
gunicorn==21.2.0